backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
mysql-connector-python
pandas
numpy
python-dotenv
